job-tracker/server/requirements.txt
# server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jobspy==1.1.79
pydantic==2.9.2
httpx==0.27.2